Move transaction manager to a split file
