+++ components/core/src/clp/ffi/ir_stream/Deserializer.cpp
@@ -14,6 +14,7 @@
 
 #include "../../ReaderInterface.hpp"
 #include "../../time_types.hpp"
+#include "../../TransactionManager.hpp"
 #include "../KeyValuePairLogEvent.hpp"
 #include "decoding_methods.hpp"
 #include "ir_unit_deserialization_methods.hpp"
@@ -22,49 +23,6 @@
 
 namespace clp::ffi::ir_stream {
 namespace {
-/**
- * Class to perform different actions depending on whether a transaction succeeds or fails. The
- * default state assumes the transaction fails.
- * @tparam SuccessHandler A cleanup lambda to call on success.
- * @tparam FailureHandler A cleanup lambda to call on failure.
- */
-template <typename SuccessHandler, typename FailureHandler>
-requires(std::is_invocable_v<SuccessHandler> && std::is_invocable_v<FailureHandler>)
-class TransactionManager {
-public:
-    // Constructor
-    TransactionManager(SuccessHandler success_handler, FailureHandler failure_handler)
-            : m_success_handler{success_handler},
-              m_failure_handler{failure_handler} {}
-
-    // Delete copy/move constructor and assignment
-    TransactionManager(TransactionManager const&) = delete;
-    TransactionManager(TransactionManager&&) = delete;
-    auto operator=(TransactionManager const&) -> TransactionManager& = delete;
-    auto operator=(TransactionManager&&) -> TransactionManager& = delete;
-
-    // Destructor
-    ~TransactionManager() {
-        if (m_success) {
-            m_success_handler();
-        } else {
-            m_failure_handler();
-        }
-    }
-
-    // Methods
-    /**
-     * Marks the transaction as successful.
-     */
-    auto mark_success() -> void { m_success = true; }
-
-private:
-    // Variables
-    SuccessHandler m_success_handler;
-    FailureHandler m_failure_handler;
-    bool m_success{false};
-};
-
 /**
  * @param tag
  * @return Whether the tag represents a schema tree node.
